third_party/WebKit/LayoutTests/fast/css/case-sensitive-attr-expected.txt

PASS window.getComputedStyle(attr1, null).color is "rgb(0, 128, 0)"
PASS window.getComputedStyle(attr2, null).color is "rgb(0, 0, 255)"
PASS window.getComputedStyle(attr3, null).color is "rgb(0, 0, 0)"
PASS window.getComputedStyle(dir1, null).color is "rgb(255, 0, 0)"
PASS window.getComputedStyle(dir2, null).color is "rgb(255, 0, 0)"
PASS successfullyParsed is true

TEST COMPLETE

This line should be green (attribute value is case sensitive).
This line should be purple (attribute value is case sensitive).
This line should be black (attribute value is case sensitive).
This line should be green (dir attribute is case insensitive).
This line should be green (dir attribute is case insensitive).
